requirements.txt
torch
torchvision
timm
transformers
gradio
mediapipe
opencv-python
numpy
pandas
Pillow
scikit-learn
tqdm
